Show property editor by double-click on label (aurusov/raox#334)
