+++ ru.bmstu.rk9.rao.ui/src/ru/bmstu/rk9/rao/ui/process/ProcessEditPart.java
@@ -8,12 +8,7 @@ import org.eclipse.gef.ConnectionEditPart;
 import org.eclipse.gef.EditPolicy;
 import org.eclipse.gef.NodeEditPart;
 import org.eclipse.gef.Request;
-import org.eclipse.gef.RequestConstants;
 import org.eclipse.gef.requests.DropRequest;
-import org.eclipse.ui.IPageLayout;
-import org.eclipse.ui.IWorkbenchPage;
-import org.eclipse.ui.PartInitException;
-import org.eclipse.ui.PlatformUI;
 
 import ru.bmstu.rk9.rao.ui.gef.EditPart;
 import ru.bmstu.rk9.rao.ui.process.connection.Connection;
@@ -35,18 +30,6 @@ public abstract class ProcessEditPart extends EditPart implements NodeEditPart {
 	@Override
 	public final List<ru.bmstu.rk9.rao.ui.gef.Node> getModelChildren() {
 		return ((ru.bmstu.rk9.rao.ui.gef.Node) getModel()).getChildren();
-	}
-
-	@Override
-	public void performRequest(Request request) {
-		if (request.getType().equals(RequestConstants.REQ_OPEN)) {
-			try {
-				IWorkbenchPage page = PlatformUI.getWorkbench().getActiveWorkbenchWindow().getActivePage();
-				page.showView(IPageLayout.ID_PROP_SHEET);
-			} catch (PartInitException e) {
-				e.printStackTrace();
-			}
-		}
 	}
 
 	@Override
